refactor: more factoring, move requests to own file
