+++ react-components/src/components/NodeCacheProvider/requests.ts
@@ -1,0 +1,107 @@
+/*!
+ * Copyright 2023 Cognite AS
+ */
+
+import { CogniteClient, CogniteInternalId, Node3D } from '@cognite/sdk';
+import { DmsUniqueIdentifier, FdmSDK, InspectResultList } from '../../utilities/FdmSDK';
+import { FdmCadEdge } from './types';
+import { INSTANCE_SPACE_3D_DATA, InModel3dEdgeProperties, SYSTEM_3D_EDGE_SOURCE } from '../../utilities/globalDataModels';
+
+export async function fetchAncestorNodesForTreeIndex(
+  modelId: number,
+  revisionId: number,
+  treeIndex: number,
+  cogniteClient: CogniteClient
+): Promise<Node3D[]> {
+  const nodeId = await treeIndexesToNodeIds(modelId, revisionId, [treeIndex], cogniteClient);
+
+  const ancestorNodes = await cogniteClient.revisions3D.list3DNodeAncestors(
+    modelId,
+    revisionId,
+    nodeId[0]
+  );
+
+  return ancestorNodes.items;
+}
+
+export async function getMappingEdgesForNodeIds(
+  modelId: number,
+  revisionId: number,
+  fdmClient: FdmSDK,
+  ancestorIds: CogniteInternalId[]
+): Promise<{ edges: FdmCadEdge[] }> {
+  const filter = {
+    and: [
+      {
+        equals: {
+          property: ['edge', 'endNode'],
+          value: {
+            space: INSTANCE_SPACE_3D_DATA,
+            externalId: `${modelId}`
+          }
+        }
+      },
+      {
+        equals: {
+          property: [
+            SYSTEM_3D_EDGE_SOURCE.space,
+            `${SYSTEM_3D_EDGE_SOURCE.externalId}/${SYSTEM_3D_EDGE_SOURCE.version}`,
+            'revisionId'
+          ],
+          value: revisionId
+        }
+      },
+      {
+        in: {
+          property: [
+            SYSTEM_3D_EDGE_SOURCE.space,
+            `${SYSTEM_3D_EDGE_SOURCE.externalId}/${SYSTEM_3D_EDGE_SOURCE.version}`,
+            'revisionNodeId'
+          ],
+          values: ancestorIds
+        }
+      }
+    ]
+  };
+
+  return await fdmClient.filterAllInstances<InModel3dEdgeProperties>(
+    filter,
+    'edge',
+    SYSTEM_3D_EDGE_SOURCE
+  );
+}
+
+export async function inspectNodes(
+  fdmClient: FdmSDK,
+  dataNodes: DmsUniqueIdentifier[]
+): Promise<InspectResultList> {
+  const inspectionResult = await fdmClient.inspectInstances({
+    inspectionOperations: { involvedViewsAndContainers: {} },
+    items: dataNodes.map((node) => ({
+      instanceType: 'node',
+      externalId: node.externalId,
+      space: node.space
+    }))
+  });
+
+  return inspectionResult;
+}
+
+export async function treeIndexesToNodeIds(
+  modelId: number,
+  revisionId: number,
+  treeIndexes: number[],
+  cogniteClient: CogniteClient
+): Promise<number[]> {
+  const outputsUrl = `${cogniteClient.getBaseUrl()}/api/v1/projects/${
+    cogniteClient.project
+  }/3d/models/${modelId}/revisions/${revisionId}/nodes/internalids/bytreeindices`;
+  const response = await cogniteClient.post<{ items: number[] }>(outputsUrl, {
+    data: { items: treeIndexes }
+  });
+  if (response.status === 200) {
+    return response.data.items;
+  } else {
+    throw Error(`treeIndex-nodeId translation failed for treeIndexes ${treeIndexes.join(',')}`);
+  }
+}
